update examples and docs
