+++ examples/experimental/json-pure-js/package.json
@@ -8,10 +8,9 @@
     "build": "NODE_ENV=production webpack --env.prod"
   },
   "dependencies": {
-    "@deck.gl/core": "^6.2.0",
-    "@deck.gl/experimental-layers": "^6.2.0",
-    "@deck.gl/json": "^6.2.0",
-    "@deck.gl/layers": "^6.2.0"
+    "@deck.gl/core": "^7.0.0-alpha",
+    "@deck.gl/json": "^7.0.0-alpha",
+    "@deck.gl/layers": "^7.0.0-alpha"
   },
   "devDependencies": {
     "buble": "^0.19.3",
