+++ examples/experimental/json-react/package.json
@@ -8,11 +8,9 @@
     "build": "NODE_ENV=production webpack --env.prod=true"
   },
   "dependencies": {
-    "@deck.gl/experimental-layers": "^6.2.0",
-    "@deck.gl/json": "^6.2.0",
+    "@deck.gl/json": "^7.0.0-alpha",
     "brace": "^0.11.1",
-    "deck.gl": "^6.2.0",
-    "luma.gl": "^6.2.0",
+    "deck.gl": "^7.0.0-alpha",
     "react": "~16.2.0",
     "react-ace": "^6.1.4",
     "react-dom": "~16.2.0",
