+++ examples/layer-browser/src/examples/index.js
@@ -1,15 +1,13 @@
 import coreLayerExamples from './core-layers';
-import experimentalLayerExamples from './experimental-layers';
+import additionalLayerExamples from './additional-layers';
 import infovisLayerExamples from './infovis-layers';
-import s2LayerExamples from './s2-layers';
 // import sampleLayerExamples from './sample-layers';
 import perfLayerExamples from './perf-layers';
 
 export default {
   ...coreLayerExamples,
-  ...experimentalLayerExamples,
+  ...additionalLayerExamples,
   ...infovisLayerExamples,
-  ...s2LayerExamples,
   // ...sampleLayerExamples,
   ...perfLayerExamples
 };
