+++ examples/website/brushing/package.json
@@ -8,7 +8,7 @@
   },
   "dependencies": {
     "d3-scale": "1.0.6",
-    "deck.gl": "^6.2.0",
+    "deck.gl": "^7.0.0-alpha",
     "react": "^16.3.0",
     "react-dom": "^16.3.0",
     "react-map-gl": "^3.3.0"
