+++ examples/website/icon/package.json
@@ -7,7 +7,7 @@
     "start": "webpack-dev-server --progress --hot --open"
   },
   "dependencies": {
-    "deck.gl": "^6.2.0",
+    "deck.gl": "^7.0.0-alpha",
     "rbush": "^2.0.1",
     "react": "^16.3.0",
     "react-dom": "^16.3.0",
