+++ examples/website/point-cloud-ply/package.json
@@ -12,7 +12,7 @@
   },
   "dependencies": {
     "d3-request": "^1.0.5",
-    "deck.gl": "^6.2.0",
+    "deck.gl": "^7.0.0-alpha",
     "gl-matrix": "^2.3.2",
     "is-little-endian": "0.0.0",
     "react": "^16.3.0",
