+++ examples/website/trips/package.json
@@ -7,8 +7,7 @@
     "start": "webpack-dev-server --progress --hot --open"
   },
   "dependencies": {
-    "deck.gl": "^6.2.0",
-    "@deck.gl/experimental-layers": "^6.2.0",
+    "deck.gl": "^7.0.0-alpha",
     "react": "^16.3.0",
     "react-dom": "^16.3.0",
     "react-map-gl": "^3.3.0"
