+++ website/contents/pages.js
@@ -383,6 +383,10 @@ export const docPages = generatePath([
             content: getDocUrl('layers/arc-layer.md')
           },
           {
+            name: 'BitmapLayer',
+            content: getDocUrl('layers/bitmap-layer.md')
+          },
+          {
             name: 'ContourLayer',
             content: getDocUrl('layers/contour-layer.md')
           },
@@ -427,6 +431,10 @@ export const docPages = generatePath([
             content: getDocUrl('layers/polygon-layer.md')
           },
           {
+            name: 'S2Layer',
+            content: getDocUrl('layers/s2-layer.md')
+          },
+          {
             name: 'ScatterplotLayer',
             content: getDocUrl('layers/scatterplot-layer.md')
           },
@@ -439,13 +447,12 @@ export const docPages = generatePath([
             content: getDocUrl('layers/text-layer.md')
           },
           {
-            name: 'S2 Layers (Experimental)',
-            children: [
-              {
-                name: 'S2Layer',
-                content: getDocUrl('api-reference/s2-layers/s2-layer.md')
-              }
-            ]
+            name: 'TileLayer',
+            content: getDocUrl('layers/tile-layer.md')
+          },
+          {
+            name: 'TripsLayer',
+            content: getDocUrl('layers/trips-layer.md')
           }
         ]
       },
